requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data
pandas>=2.0.0
pybaseball>=2.2.5
requests>=2.31.0

# Scheduling / utils
python-dotenv>=1.0.0
